+++ NEWS.md
@@ -1,3 +1,13 @@
+# openxlsx2 (in development)
+
+## Breaking changes
+
+* Conditional style defaults for `create_dxfs_style()` have changed to be more permissive. Previously we shipped a default font, default font size and font color. This has been changed to better reflect a behavior the user expects. [343](https://github.com/JanMarvin/openxlsx2/issues/343)
+
+
+***************************************************************************
+
+
 # openxlsx2 0.3
 
 ## New features